+++ pads/src/com/mynetpcb/pad/shape/Circle.java
@@ -217,15 +217,15 @@ public class Circle  extends Shape implements ArcGerberable,Resizeable,Externali
         Element  element= (Element)node;
         this.setCopper(Layer.Copper.valueOf(element.getAttribute("copper")));    
              
-        int xx=(Integer.parseInt(element.getAttribute("x")));
-        int yy=(Integer.parseInt(element.getAttribute("y")));        
+        double xx=(Double.parseDouble(element.getAttribute("x")));
+        double yy=(Double.parseDouble(element.getAttribute("y")));        
 
         if(element.getAttribute("width").length()>0){  
             int diameter=(Integer.parseInt(element.getAttribute("width")));
             this.circle.pc.set(xx+((diameter/2)),yy+((diameter/2)));
             this.circle.r=(diameter/2); 
         }else{
-            int radius=(Integer.parseInt(element.getAttribute("radius"))); 
+            double radius=(Double.parseDouble(element.getAttribute("radius"))); 
             this.circle.pc.set(xx,yy);
             this.circle.r=radius;   
         }
